// src/types/formularioDeAvaliacaoCursoAprofundamentoMindfulness.d.ts
export interface FormularioDeAvaliacaoCursoAprofundamentoMindfulnessI {
  /** Auto-incrementing primary key */
  id: number;

  createdAt?: string;

  fullName?: string;
  birthDate?: string;
  satisfactionLevel?: string;

  likedMostAndHighlights?: string;
  likedLeastSuggestions?: string;
  personalChange?: string;
  selfDifference?: string;
  knowledgeMomentsImportance?: string;
  tellAFriend?: string;
  benefitsBeyondPrevious?: string;
  personalPracticeStimulation?: string;
  authorizeUse?: string;
}